requirements.txt
python-dotenv
langchain
langchain-community
langchain-ollama
langchain-text-splitters
langgraph
sentence-transformers
chromadb
pypdf
